Disabling of c.j minor tweaks to c.lw and c.sw tests.
Changes for c.j include the usual updates to the tablegen
sources. This time I need to patch the lld relaxation code to disable
a call to c.j relaxation. The effectiveness of this relies on the
correctness of the .riscv.attributes section.

Correct typos in the clw.ll and csw.ll tests.