Disabled the driver's check that disallows -emit-llvm when linking
Disabled this check to allow the -emit-llvm switch to be used at link
time. This is to work around picolibc's meson build system using the
switch at configure time. It appears to use the compiler switches for
linking even when the linker flags have been explicitly set. The
result was that the compiler appeared to lack support for features
such as __builtin_alloc() that _are_ available.