Add disable of BGEU. Split the pass out. Add extra predicates.
Disabling of the BGEU instruction is not supported. Tests exposed a
problem whereby the atomic-expand pass was adding branch instructions
that may be disabled. This was solved by splitting the Keysom
expansion pass out of the PreRA expand pass (this needed to happen
anyway) and running it twise: before RA and very late after atomic
expansion.