Ensure that endian-ness is correct. Fix stack spilling.
Setting the correct portions of halfwords and words has proven to be
tricky in the absence of SB or SH instructions. Disabling any of the
byte load or store or instructions will cause the size of spilled
bytes to be two bytes; disabling halfword instructions will increase
the size to four bytes.

Added assertions to ensure that load/store expansion doesn't ever
happen after register allocation.

TODO: optimized odd/even handling for halfwords in the same way as I
have to bytes. Improved expansion of byte loads/stores when *both*
B&H instructions are disabled (right now the expansion works but is
inefficient).