Rework LB/LBU, LH/LHU. Remove SB/SH.
Reworked the LB/LHU support so that at most one of these instructions
can be disabled at a time. When one is disabled, its replacement code
is written in terms of the other. The same applies to LH/LHU. Support
for disabling SB/SH removed altogher. Changes to stack spilling are
removed.